fix: satisfy backend file-size gate after main rebase
Move session runtime build structs into factory_build.go and relocate
packaged invocation helpers out of subsystem_transitioner.go.
