+++ pkg/factory/subsystems/subsystem_history.go
@@ -2,11 +2,14 @@ package subsystems
 
 import (
 	"context"
+	"fmt"
 	"strings"
 
 	"github.com/portpowered/infinite-you/pkg/factory/state"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/logging"
+	"github.com/portpowered/infinite-you/pkg/packagedfactories/goal"
+	"github.com/portpowered/infinite-you/pkg/packagedfactories/tts"
 	"github.com/portpowered/infinite-you/pkg/petri"
 )
 
@@ -115,4 +118,85 @@ func evaluateStopWords(stopWords []string, output string) interfaces.WorkOutcome
 		}
 	}
 	return interfaces.OutcomeFailed
+}
+
+func firstNonResourceInput(inputs []interfaces.TokenColor) *interfaces.TokenColor {
+	for i := range inputs {
+		if inputs[i].DataType != interfaces.DataTypeResource && inputs[i].WorkTypeID != interfaces.SystemTimeWorkTypeID {
+			return &inputs[i]
+		}
+	}
+	for i := range inputs {
+		if inputs[i].DataType != interfaces.DataTypeResource {
+			return &inputs[i]
+		}
+	}
+	return nil
+}
+
+func applyPackagedTTSInvocationMetadata(
+	token *interfaces.Token,
+	workstation *interfaces.FactoryWorkstationConfig,
+	workerOutput string,
+	inputColors []interfaces.TokenColor,
+	runtimeConfig interfaces.RuntimeWorkstationLookup,
+) error {
+	if token == nil || !tts.ShouldFormatInvocationMetadata(workstation) {
+		return nil
+	}
+
+	traceID := ""
+	if source := firstNonResourceInput(inputColors); source != nil {
+		traceID = strings.TrimSpace(source.TraceID)
+	}
+
+	backendLabel := ""
+	if workstation != nil && runtimeConfig != nil {
+		if lookup, ok := runtimeConfig.(interfaces.RuntimeDefinitionLookup); ok {
+			if worker, ok := lookup.Worker(strings.TrimSpace(workstation.WorkerTypeName)); ok && worker != nil {
+				backendLabel = tts.BackendLabelFromWorker(worker)
+			}
+		}
+	}
+
+	metadataContent, err := tts.MetadataContentFromWorkerOutput(workerOutput, traceID, "", backendLabel)
+	if err != nil {
+		return fmt.Errorf("shape packaged tts invocation metadata: %w", err)
+	}
+
+	token.Color.Content = metadataContent
+	token.Color.Payload = nil
+	return nil
+}
+
+func applyPackagedGoalInvocationSummary(
+	token *interfaces.Token,
+	workstation *interfaces.FactoryWorkstationConfig,
+	workerOutput string,
+	runtimeConfig interfaces.RuntimeWorkstationLookup,
+) error {
+	if token == nil || !goal.ShouldFormatInvocationSummary(workstation) {
+		return nil
+	}
+	if strings.TrimSpace(workerOutput) == "" {
+		return nil
+	}
+
+	stopToken := ""
+	if workstation != nil && runtimeConfig != nil {
+		if lookup, ok := runtimeConfig.(interfaces.RuntimeDefinitionLookup); ok {
+			if worker, ok := lookup.Worker(strings.TrimSpace(workstation.WorkerTypeName)); ok && worker != nil {
+				stopToken = strings.TrimSpace(worker.StopToken)
+			}
+		}
+	}
+
+	summaryContent, err := goal.SummaryContentFromWorkerOutput(workerOutput, stopToken)
+	if err != nil {
+		return fmt.Errorf("shape packaged goal invocation summary: %w", err)
+	}
+
+	token.Color.Content = summaryContent
+	token.Color.Payload = nil
+	return nil
 }
